get caching back together; update changelog
