+++ tests/test_model_cache_store.py
@@ -1,5 +1,7 @@
 from toshi_hazard_store import model
-from toshi_hazard_store.v2.db_adapter.sqlite import sqlite_store as cache_store
+
+# from toshi_hazard_store.v2.db_adapter.sqlite import sqlite_store as cache_store
+from toshi_hazard_store.model.caching import cache_store
 
 
 class TestCacheStoreSQLExpressions:
